requirements.txt
# FastAPI y servidor
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Base de datos
sqlalchemy==2.0.36
alembic==1.14.0

# Autenticación y seguridad
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# Validación de datos
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.27.2
faker==33.1.0

# Utilidades
python-dotenv==1.0.1

